Day 24 part 2
I got to this by reading how other people did it. A common theme was assuming that the velocities would be small and brute forcing it.

If we know the start velocity, we have three unknowns (x, y, z) and 1 unknown per line (t_i). So if we consider two lines, we get 5 unknowns and 6 equations. I solved for the 5 unknowns by representing them as a matrix multipliciation and inverting the matrix. For a given velocity, we compare the solutions for two different lines. The error between those should be very small for the right solution, so we print it out every time the error goes down.